@@ -14,10 +14,10 @@
 		}
 	],
 	"require": {
-		"php": ">=5.3.3"
+		"php": ">=7.1"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "~4.0",
+		"phpunit/phpunit": "~8.0",
 		"overtrue/phplint": "^1.2",
 		"squizlabs/php_codesniffer": "^3.5"
 	},
