@@ -2,11 +2,11 @@
 
 require_once __DIR__ . "/../lib/PostHog/Client.php";
 
-class ConsumerLibCurlTest extends PHPUnit_Framework_TestCase
+class ConsumerLibCurlTest extends PHPUnit\Framework\TestCase
 {
   private $client;
 
-  public function setUp()
+  public function setUp(): void
   {
     date_default_timezone_set("UTC");
     $this->client = new PostHog_Client(
